requirements.txt
Flask
Flask-Login
Flask-WTF
Flask-SQLAlchemy
Flask-Migrate
PyMySQL
Werkzeug
Pillow
torch
torchvision
pandas
requests
gunicorn

